betapp/server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
